requirements.txt
Django==3.2.15
Pillow==9.2.0
python-dotenv==0.21.0
requests==2.28.1
simplejson==3.17.6
django-filter==22.1